… docs/database-diagram.jpg