## Context

The Courses::Query service was experiencing high latency from the DB (from 163 to 200 ms) for searches due to:

* Duplicate provider JOINs
* Unindexed filters on course_site and course
* Disk-based sorting and sequential scans in the join

Changes

* Added idx_course_site_published_filtered to speed up course_site filtering by status and publish flags.
* Added idx_course_provider_active to cover provider→course joins and application_status filtering.
* Introduced idx_course_ordering_optimized to provide a covering index for ORDER BY operations, eliminating in-memory sorts.
* Added idx_provider_recruitment_cycle_optimized to accelerate provider lookup by recruitment_cycle_id and discarded_at.

## Before (Local)

Normal query: between 160 - 230 ms (each - query and counting)

## After (Local)

Normal query: 15 to 20 ms (each - query and counting)

## Query Plan Improvements

Before: External merge sort using 997 temp disk blocks
After: Quicksort in memory using 768KB

Before: Sequential scan filtering 4,973 provider rows
After: Index-optimized scans with minimal filtering

Before: 19,683 intermediate rows processed
After: Efficient join order with reduced intermediate processing

### K6 load testing (against local)

Baseline scenario:

**Total Requests**: 14,334 (5.7% more requests handled!)
**Failed Requests**: 0 (100% success rate maintained)
**Failed Checks**: 0 (functional tests passed)

| Metric | Before | After | Improvement | Status |
|--------|--------|-------|-------------|--------|
| **Average** | 1,723ms | **1,520ms** | **-203ms (-12%)** | 🟡 **Better** |
| **Maximum** | 4,142ms | **3,477ms** | **-665ms (-16%)** | 🟡 **Better** |
| **Median** | 1,669ms | **1,499ms** | **-170ms (-10%)** | 🟡 **Better** |
| **Minimum** | 48ms | **38ms** | **-10ms (-21%)** | ✅ **Excellent** |
| **90th Percentile** | 3,049ms | **2,601ms** | **-448ms (-15%)** | 🟡 **Better** |
| **95th Percentile** | 3,454ms | **2,790ms** | **-664ms (-19%)** | ✅ **Green!** |

1. 95th Percentile Breakthrough! 🎊
Before: 3,454ms (🔴 RED - breached threshold)
After: 2,790ms (✅ GREEN - under threshold!)

Improvement: 19% faster for worst-case users

2. Consistent Performance Gains Across All Metrics

Every single metric improved by 10-21%
No performance regressions anywhere
Maintained 100% success rate

3. Better Throughput

Before: 13,556 requests
After: 14,334 requests (+778 more requests, +5.7%)